backend/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.0
ecdsa==0.18.0
fastapi==0.109.0
fastapi-jwt-auth==0.5.0
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
kombu==5.3.5
passlib==1.7.4
prompt-toolkit==3.0.43
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
uvicorn==0.26.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
zope.event==5.0
zope.interface==6.2
